# service/course-enrollment-service/openapi/api.yaml
openapi: 3.1.0

info:
  title: Course Enrollment API
  version: 1.0.0
  description: REST API for course enrollment system
servers:
  - url: /
tags:
  - name: courses
    description: Course management
paths:
  /courses:
    get:
      tags:
        - courses
      operationId: getCourses
      summary: List courses with pagination
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: Paginated list of courses
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoursePageResponse'
        '400':
          description: Invalid pagination parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  parameters:
    PageParam:
      name: page
      in: query
      description: Zero-based page index
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    SizeParam:
      name: size
      in: query
      description: Number of items per page
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortParam:
      name: sort
      in: query
      description: "Sort field and direction, e.g. title,asc"
      required: false
      schema:
        type: string
  schemas:
    CourseDto:
      type: object
      required:
        - id
        - code
        - title
        - credits
        - status
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
          example: CS101
        title:
          type: string
          example: Introduction to Computer Science
        description:
          type: string
        credits:
          type: integer
          minimum: 1
        capacity:
          type: integer
          minimum: 1
        enrolledCount:
          type: integer
          minimum: 0
        status:
          $ref: '#/components/schemas/CourseStatusDto'
    CourseStatusDto:
      type: string
      enum:
        - OPEN
        - CLOSED
        - CANCELLED
    PaginationMetaDto:
      type: object
      required:
        - page
        - size
        - totalElements
        - totalPages
      properties:
        page:
          type: integer
          description: Current zero-based page index
        size:
          type: integer
          description: Page size requested
        totalElements:
          type: integer
          format: int64
          description: Total number of courses
        totalPages:
          type: integer
          description: Total number of pages
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
    CoursePageResponse:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/CourseDto'
        pagination:
          $ref: '#/components/schemas/PaginationMetaDto'
    ErrorResponse:
      description: Problem details per RFC 9457
      type: object
      properties:
        type:
          type: string
          format: uri
          description: URI identifying the problem type
          default: about:blank
        title:
          type: string
          description: Short human-readable summary of the problem type
        status:
          type: integer
          description: HTTP status code
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence
        instance:
          type: string
          format: uri
          description: URI identifying the specific occurrence