# service/course-enrollment-service/src/generated/openapi/src/main/java/com/tranvodev/courseenrollment/api/CoursesApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.22.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.tranvodev.courseenrollment.api;

import com.tranvodev.courseenrollment.model.dto.CoursePageResponse;
import com.tranvodev.courseenrollment.model.dto.ErrorResponse;
import org.springframework.lang.Nullable;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.22.0")
@Validated
@Tag(name = "courses", description = "Course management")
public interface CoursesApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_GET_COURSES = "/courses";
    /**
     * GET /courses : List courses with pagination
     *
     * @param page Zero-based page index (optional, default to 0)
     * @param size Number of items per page (optional, default to 20)
     * @param sort Sort field and direction, e.g. title,asc (optional)
     * @return Paginated list of courses (status code 200)
     *         or Invalid pagination parameters (status code 400)
     */
    @Operation(
        operationId = "getCourses",
        summary = "List courses with pagination",
        tags = { "courses" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Paginated list of courses", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CoursePageResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid pagination parameters", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = CoursesApi.PATH_GET_COURSES,
        produces = { "application/json" }
    )
    default ResponseEntity<CoursePageResponse> getCourses(
        @Min(value = 0) @Parameter(name = "page", description = "Zero-based page index", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false, defaultValue = "0") Integer page,
        @Min(value = 1) @Max(value = 100) @Parameter(name = "size", description = "Number of items per page", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false, defaultValue = "20") Integer size,
        @Parameter(name = "sort", description = "Sort field and direction, e.g. title,asc", in = ParameterIn.QUERY) @Valid @RequestParam(value = "sort", required = false) @Nullable String sort
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"pagination\" : { \"size\" : 2, \"totalPages\" : 9, \"hasPrevious\" : true, \"hasNext\" : true, \"page\" : 5, \"totalElements\" : 7 }, \"data\" : [ { \"code\" : \"CS101\", \"credits\" : 1, \"enrolledCount\" : 0, \"description\" : \"description\", \"id\" : 0, \"title\" : \"Introduction to Computer Science\", \"capacity\" : 1, \"status\" : \"OPEN\" }, { \"code\" : \"CS101\", \"credits\" : 1, \"enrolledCount\" : 0, \"description\" : \"description\", \"id\" : 0, \"title\" : \"Introduction to Computer Science\", \"capacity\" : 1, \"status\" : \"OPEN\" } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"instance\" : \"https://openapi-generator.tech\", \"detail\" : \"detail\", \"type\" : \"https://openapi-generator.tech\", \"title\" : \"title\", \"status\" : 3 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}